Pull up OperationProcessorTester
