+++ vespa-http-client/src/test/java/com/yahoo/vespa/http/client/core/operationProcessor/IncompleteResultsThrottlerTest.java
@@ -13,9 +13,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Random;
 
-import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.anyDouble;
 import static org.mockito.ArgumentMatchers.eq;
@@ -27,14 +25,14 @@ public class IncompleteResultsThrottlerTest {
     @Test
     public void simpleStaticQueueSizeTest() {
         IncompleteResultsThrottler incompleteResultsThrottler = new IncompleteResultsThrottler(2, 2, null, null);
-        assertThat(incompleteResultsThrottler.waitingThreads(), is(0));
+        assertEquals(0, incompleteResultsThrottler.waitingThreads());
         incompleteResultsThrottler.operationStart();
         incompleteResultsThrottler.operationStart();
-        assertThat(incompleteResultsThrottler.waitingThreads(), is(2));
+        assertEquals(2, incompleteResultsThrottler.waitingThreads());
         incompleteResultsThrottler.resultReady(true);
-        assertThat(incompleteResultsThrottler.waitingThreads(), is(1));
+        assertEquals(1, incompleteResultsThrottler.waitingThreads());
         incompleteResultsThrottler.resultReady(true);
-        assertThat(incompleteResultsThrottler.waitingThreads(), is(0));
+        assertEquals(0, incompleteResultsThrottler.waitingThreads());
     }
 
     /**
